requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
requests==2.28.2